Change name, narrow types
