+++ src/components/index.ts
@@ -149,6 +149,7 @@ export { default as FlowRunTimelineTaskDetails } from './FlowRunTimelineTaskDeta
 export { default as FlowsDeleteButton } from './FlowsDeleteButton.vue'
 export { default as FlowsFilterGroup } from './FlowsFilterGroup.vue'
 export { default as FlowsPageEmptyState } from './FlowsPageEmptyState.vue'
+export { default as FormattedDate } from './FormattedDate.vue'
 export { default as FourZeroFour } from './FourZeroFour.vue'
 export { default as IconTextCount } from './IconTextCount.vue'
 export { default as IntervalScheduleForm } from './IntervalScheduleForm.vue'
@@ -217,7 +218,6 @@ export { default as ParametersTable } from './ParametersTable.vue'
 export { default as ProcessTypeBadge } from './ProcessTypeBadge.vue'
 export { default as QuickRunParametersModal } from './QuickRunParametersModal.vue'
 export { default as ResultsCount } from './ResultsCount.vue'
-export { default as RichTimestamp } from './RichTimestamp.vue'
 export { default as RunMenu } from './RunMenu.vue'
 export { default as SavedFilters } from './SavedFilters.vue'
 export { default as SavedFiltersDeleteModal } from './SavedFiltersDeleteModal.vue'
